check for element before starting puppeteer

+++ service.js
@@ -3,6 +3,15 @@ const fs = require('fs')
 
 const service = async () => {
     const url = 'https://orf.at/'
+
+    const { default: fetch } = await import('node-fetch')
+    const request = await fetch(url)
+    const response = await request.text()
+
+    const check = response.includes('oon-grid-item-version-4k')
+    if (check) { return null }
+
+
     const path = `${__dirname}/screenshot.jpg`
     if (fs.existsSync(path)) { fs.unlinkSync(path) }
 
@@ -14,12 +23,6 @@ const service = async () => {
     const page = await browser.newPage()
     await page.setJavaScriptEnabled(false)
     await page.goto(url);
-
-    const check = await page.evaluate(() => {
-        return !document.querySelector('.oon-grid-item-version-4k')
-    });
-
-    if (!check) { return null }
     await page.screenshot({ path })
     return path
 }
